snippets/requirements.txt
### include any additional libraries used in your code here
git+https://github.com/skhatiri/Aerialist.git
pymoo==0.6.1.1
numpy==1.23.2
similaritymeasures==0.5.0
matplotlib==3.6.1
pyulog==0.9.0
mavsdk==1.4.1
keyboard==0.13.5
python-decouple==3.6
csv-logger==1.3.0
statsmodels==0.13.2
webdavclient3==3.14.6
ruptures==1.1.7
shapely==1.8.4
validators==0.20.0
pyyaml==6.0
munch==2.5.0
rospkg==1.4.0
defusedxml==0.7.1
tslearn==0.5.2
h5py==3.9.0
